requirements.txt
streamlit==1.15.1
plotly==5.10.10
pandas==1.4.3
numpy==1.23.1
foluim==0.13.0
matplotlip-inline==0.1.6
haversine==2.7.0
streamlit-folium==0.7.0
Pillow==9.2.0